requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
iniconfig==2.3.0
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.1
patsy==1.0.2
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
statsmodels==0.14.6
threadpoolctl==3.6.0
